Fix label creation for multiple subtasks
Fixed a small bug in :GInsert behavior. In a project with many tasks
which each have multiple subtasks, the p:proj:task:subtask tag isn't
always created correctly due to the project/task tree being traversed
incorrectly.

This fix ensures that the project/task tree is always traversed in the
direction of the root of the tree.